mainapp/filters.py
from django_filters import rest_framework as filters
from .models import Transaction


class TransactionFilter(filters.FilterSet):
    min_amount = filters.NumberFilter(field_name="amount", lookup_expr='gte')
    max_amount = filters.NumberFilter(field_name="amount", lookup_expr='lte')
    min_date = filters.DateFilter(field_name="date", lookup_expr='gte')
    max_date = filters.DateFilter(field_name="date", lookup_expr='lte')


    class Meta:
        model = Transaction
        fields = ['type', 'min_amount', 'max_amount', 'min_date', 'max_date']
